[agent] refactor(ner): split ner into focused submodules
